refactor: Update package structure for remote data source
