+++ data/src/main/java/com/london/data/remote/source/reviews/ReviewsRemoteDataSourceImpl.kt
@@ -1,9 +1,9 @@
 package com.london.data.remote.source.reviews
 
-import com.london.data.datasource.remote.ApiResponse
-import com.london.data.datasource.remote.BaseRemoteDatasource
-import com.london.data.datasource.remote.reviews.api.ReviewsApiService
+import com.london.data.remote.model.ApiResponse
 import com.london.data.remote.model.reviews.model.ReviewResponse
+import com.london.data.remote.service.reviews.ReviewsApiService
+import com.london.data.remote.source.base.BaseRemoteDatasource
 import org.koin.core.annotation.Single
 
 @Single
