+++ data/src/main/java/com/london/data/remote/source/search/SearchRemoteDataSource.kt
@@ -2,10 +2,10 @@
 
 package com.london.data.remote.source.search
 
-import com.london.data.datasource.remote.ApiResponse
-import com.london.data.datasource.remote.search.model.SearchActorRemote
-import com.london.data.datasource.remote.search.model.SearchMovieRemote
-import com.london.data.datasource.remote.search.model.SearchTvShowRemote
+import com.london.data.remote.model.ApiResponse
+import com.london.data.remote.model.search.model.SearchActorRemote
+import com.london.data.remote.model.search.model.SearchMovieRemote
+import com.london.data.remote.model.search.model.SearchTvShowRemote
 import com.london.domain.KoverIgnore
 
 interface SearchRemoteDataSource {
